XML descriptor for composite component (2707)
